Modiffied the AuthorController so it returns 404 when a user or author id is not found.

+++ src/main/java/com/library/library/controller/AuthorController.java
@@ -20,9 +20,15 @@ public class AuthorController {
 
     @Autowired
     private AuthorService authorService;
+    @Autowired
+    private UserController userService;
 
     @PostMapping("/users/{userId}/authors")
     public ResponseEntity<Author> createAuthor(@PathVariable(value = "userId") Long userId, @RequestBody Author authorRequest) {
+        Optional<User> user = Optional.ofNullable(userService.getUserById(userId).getBody());
+        if (!user.isPresent()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         Author author = authorService.createAuthor(userId, authorRequest);
         return new ResponseEntity<>(author, HttpStatus.CREATED);
     }
@@ -30,11 +36,18 @@ public class AuthorController {
     @GetMapping("/authors/{id}")
     public ResponseEntity<Author> getAuthorsById(@PathVariable(value = "id") Long id) {
         Author author = authorService.getAuthorById(id);
+        if (author == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         return new ResponseEntity<>(author, HttpStatus.OK);
     }
 
     @GetMapping("/users/{userId}/authors")
     public ResponseEntity<List<Author>> getAllAuthorsByUserId(@PathVariable(value = "userId") Long userId) {
+        Optional<User> user = Optional.ofNullable(userService.getUserById(userId).getBody());
+        if (!user.isPresent()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         List<Author> authors = authorService.getAllAuthorsByUserId(userId);
         return new ResponseEntity<>(authors, HttpStatus.OK);
     }
@@ -42,17 +55,28 @@ public class AuthorController {
     @PutMapping("/authors/{id}")
     public ResponseEntity<Author> updateAuthor(@PathVariable("id") long id, @RequestBody Author authorRequest) {
         Author author = authorService.updateAuthor(id, authorRequest);
+        if (author == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         return new ResponseEntity<>(author, HttpStatus.OK);
     }
 
     @DeleteMapping("/authors/{id}")
     public ResponseEntity<Void> deleteAuthor(@PathVariable("id") long id) {
+        Author author = authorService.getAuthorById(id);
+        if (author == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         authorService.deleteAuthor(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/users/{userId}/authors")
     public ResponseEntity<Void> deleteAllAuthorsOfUser(@PathVariable(value = "userId") Long userId) {
+        Optional<User> user = Optional.ofNullable(userService.getUserById(userId).getBody());
+        if (!user.isPresent()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         authorService.deleteAllAuthorsOfUser(userId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
